+++ src/common/utils/utils.ts
@@ -1,7 +1,6 @@
 import { createHash, createHmac, randomInt } from "crypto";
-import rimraf = require("rimraf");
+import { rimraf } from "rimraf";
 import { parse } from "url";
-import { promisify } from "util";
 import StorageErrorFactory from "../../blob/errors/StorageErrorFactory";
 import { VALID_CSHARP_IDENTIFIER_REGEX } from "./constants";
 
@@ -9,7 +8,7 @@ import { VALID_CSHARP_IDENTIFIER_REGEX } from "./constants";
 // tslint:disable-next-line:no-var-requires
 export const lfsa = require("lokijs/src/loki-fs-structured-adapter.js");
 
-export const rimrafAsync = promisify(rimraf);
+export const rimrafAsync = rimraf;
 
 export function minDate(date1: Date, date2: Date): Date {
   return date1 > date2 ? date2 : date1;